cart_total_price_with_stp_logic

+++ app/src/main/java/vn/quanprolazer/fashione/utilities/Util.kt
@@ -119,7 +119,7 @@ fun convertPriceStringToCurrencyString(text: String, currencyCode: String = "VND
     val format: NumberFormat = NumberFormat.getCurrencyInstance()
     format.maximumFractionDigits = 0
     format.currency = Currency.getInstance(currencyCode)
-    return format.format(text.toInt())
+    return format.format(text.toFloat())
 }
 
 /**
